Commit #11:  Event system should be fully operational for tile based textbox events.  TODO: Add projectiles and support for other tile based collision events!

+++ gameEngine/player.hpp
@@ -78,8 +78,8 @@ private:
     int speed;
     
     // Player's movement position
-    int xVelocity;
-    int yVelocity;
+    double xVelocity;
+    double yVelocity;
     
     // Bools for which axis the player is moving on
     bool movingX, movingY;
